Title: 사각형 채우기 3, Time: 0ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 260421/사각형 채우기 3/rectangle-fill-3.java
@@ -5,7 +5,7 @@ public class Main {
         int n = sc.nextInt();
         // Please write your code here.
 
-        int[] dp = new int[1001];
+        long[] dp = new long[1001];
         dp[0] = 1;
         dp[1] = 2;
         dp[2] = 7;
